Renaming model/view files
