+++ src/snooker_ball_tracker/models/logging/__init__.py
@@ -1,3 +1,3 @@
-from .ball_count_model import BallCountModel
-from .snapshot_model import SnapshotModel
-from .balls_potted_list_model import BallsPottedListModel
+from .ball_count import BallCountModel
+from .snapshot import SnapshotModel
+from .balls_potted_list import BallsPottedListModel
