+++ src/snooker_ball_tracker/models/settings/ball_detection_setting.py
