+++ src/snooker_ball_tracker/models/settings/hsv_colour.py
